backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.2
anthropic==0.25.0
python-dotenv==1.0.0
pydantic==2.5.0
yfinance>=1.2.0
PyMuPDF==1.23.7
pytz>=2024.1
reportlab==4.0.7
google-generativeai==0.5.4
